+++ XBMC Remote/ShowInfoViewController.m
@@ -469,7 +469,7 @@ int count=0;
         NSString *title;
         for (int i = 0; i < numActions; i++) {
              title = [sheetActions objectAtIndex:i];
-            if ([title isEqualToString:NSLocalizedString(@"Record", nil)] && isRecording.alpha == 1.0f) {
+            if ([title isEqualToString:NSLocalizedString(@"Record", nil)] && isRecording.alpha == 1.0) {
                 title = NSLocalizedString(@"Stop Recording", nil);
             }
             [actionSheetView addButtonWithTitle:title];
@@ -519,14 +519,14 @@ int count=0;
                      animations: ^{
                          CGRect frame;
                          frame = voteLabel.frame;
-                         if (isRecording.alpha == 0.0f) {
-                             isRecording.alpha = 1.0f;
+                         if (isRecording.alpha == 0.0) {
+                             isRecording.alpha = 1.0;
                              frame.origin.x += dotSize + dotSizePadding;
                              frame.size.width -= dotSize + dotSizePadding;
                              [voteLabel setFrame:frame];
                          }
                          else {
-                             isRecording.alpha = 0.0f;
+                             isRecording.alpha = 0.0;
                              frame.origin.x -= dotSize + dotSizePadding;
                              frame.size.width += dotSize + dotSizePadding;
                              [voteLabel setFrame:frame];
@@ -741,7 +741,7 @@ int h=0;
     [activityIndicatorView stopAnimating];
     jewelView.alpha = 0;
     jewelView.image = image;
-    [self alphaImage:jewelView AnimDuration:0.1 Alpha:1.0f];
+    [self alphaImage:jewelView AnimDuration:0.1 Alpha:1.0];
 }
 
 -(void)setIOS7barTintColor:(UIColor *)tintColor{
@@ -1216,11 +1216,11 @@ int h=0;
             isRecording = [[UIImageView alloc] initWithFrame:CGRectMake(frame.origin.x, frame.origin.y + (frame.size.height/2 - dotSize/2), dotSize, dotSize)];
             [isRecording setImage:[UIImage imageNamed:@"button_timer.png"]];
             [isRecording setContentMode:UIViewContentModeScaleAspectFill];
-            isRecording.alpha = 0.0f;
+            isRecording.alpha = 0.0;
             [isRecording setBackgroundColor:[UIColor clearColor]];
             [scrollView addSubview:isRecording];
             if ([[item objectForKey:@"hastimer"] boolValue] == YES) {
-                isRecording.alpha = 1.0f;
+                isRecording.alpha = 1.0;
                 frame.origin.x += dotSize + dotSizePadding;
                 frame.size.width -= dotSize + dotSizePadding;
                 [voteLabel setFrame:frame];
@@ -1663,7 +1663,7 @@ int h=0;
         [self alphaView:closeButton AnimDuration:1.5 Alpha:0];
         [self alphaView:scrollView AnimDuration:1.5 Alpha:1];
         if (!enableKenBurns) {
-            [self alphaImage:fanartView AnimDuration:1.5 Alpha:0.2f];// cool
+            [self alphaImage:fanartView AnimDuration:1.5 Alpha:0.2];// cool
         }
         else {
             [self alphaView:self.kenView AnimDuration:1.5 Alpha:0.2];// cool
@@ -1708,7 +1708,7 @@ int h=0;
                              completion:^(BOOL finished) {}
              ];
             if (self.kenView != nil) {
-                float alphaValue = 1;
+                CGFloat alphaValue = 1;
                 [UIView animateWithDuration:0.2
                                  animations:^ {
                                      self.kenView.alpha = 0;
@@ -1779,7 +1779,7 @@ int h=0;
     }
 }
 
--(void)alphaImage:(UIImageView *)image AnimDuration:(float)seconds Alpha:(float)alphavalue{
+-(void)alphaImage:(UIImageView *)image AnimDuration:(float)seconds Alpha:(CGFloat)alphavalue{
     [UIView beginAnimations:nil context:nil];
 	[UIView setAnimationDuration:seconds];
     [UIView setAnimationCurve:UIViewAnimationCurveEaseInOut];
@@ -1790,7 +1790,7 @@ int h=0;
     [UIView commitAnimations];
 }
 
--(void)alphaView:(UIView *)view AnimDuration:(float)seconds Alpha:(float)alphavalue{
+-(void)alphaView:(UIView *)view AnimDuration:(float)seconds Alpha:(CGFloat)alphavalue{
     [UIView beginAnimations:nil context:nil];
 	[UIView setAnimationDuration:seconds];
     [UIView setAnimationCurve:UIViewAnimationCurveEaseInOut];
@@ -1836,7 +1836,7 @@ int h=0;
 - (void)tableView:(UITableView *)tableView willDisplayCell:(UITableViewCell *)cell forRowAtIndexPath:(NSIndexPath *)indexPath {
     if ([AppDelegate instance].serverVersion > 11  && ![self isModal]) {
         cell.accessoryView = [[UIImageView alloc] initWithImage:[UIImage imageNamed:@"table_arrow_right_selected.png"]];
-        cell.accessoryView.alpha = 0.5f;
+        cell.accessoryView.alpha = 0.5;
     }
     else {
         [cell setSelectionStyle:UITableViewCellSelectionStyleNone];
@@ -2101,7 +2101,7 @@ int h=0;
         self.navigationController.navigationBar.tintColor = TINT_COLOR;
         toolbar.tintColor = TINT_COLOR;
     }
-    float alphaValue = 0.2;
+    CGFloat alphaValue = 0.2;
     if (closeButton.alpha==1){
         alphaValue = 1;
         [self.navigationController setNavigationBarHidden:YES animated:YES];
@@ -2132,7 +2132,7 @@ int h=0;
 
 -(void)viewDidDisappear:(BOOL)animated{
     [super viewDidDisappear:animated];
-    [self alphaImage:fanartView AnimDuration:0.3 Alpha:0.0f];
+    [self alphaImage:fanartView AnimDuration:0.3 Alpha:0.0];
     if (self.kenView != nil){
         [UIView animateWithDuration:0.3
                          animations:^{
@@ -2221,7 +2221,7 @@ int h=0;
 
 -(void)willRotateToInterfaceOrientation:(UIInterfaceOrientation)toInterfaceOrientation duration:(NSTimeInterval)duration{
     if (self.kenView != nil){
-        float alphaValue = 0.2;
+        CGFloat alphaValue = 0.2;
         if (closeButton.alpha==1){
             alphaValue = 1;
         }
